requirements.txt
Flask
transformers
torch
sentencepiece
tqdm
requests
google-api-python-client
python-dotenv
lxml[html_clean]
google-generativeai
newspaper3k


